- Correct handling of termination of processes from the process   manager. - Make OCR processor quiet - Remove FFFE byte order marks from html generated by pdftohtml   before processing - Remove html and text files after processing them from the queue

+++ scrapy-webscanner/scanner/processors/pdf.py
@@ -2,6 +2,7 @@ from processor import Processor
 import shutil
 import os
 import subprocess
+import regex
 
 
 class PDFProcessor(Processor):
@@ -15,11 +16,24 @@ class PDFProcessor(Processor):
 
     def convert(self, item, tmp_dir):
         # Move file to temp dir before conversion
-        shutil.move(item.file_path, tmp_dir)
         new_file_path = os.path.join(tmp_dir, os.path.basename(item.file_path))
-        return_code = subprocess.call(
-            ["pdftohtml", "-noframes", "-hidden", "-q", new_file_path]
-        )
+        if item.file_path != new_file_path:
+            shutil.move(item.file_path, tmp_dir)
+        return_code = subprocess.call([
+            "pdftohtml", "-noframes", "-hidden", "-enc", "UTF-8",
+            "-q", new_file_path
+        ])
+
+        if return_code != 0:
+            return False
+
+        # Have to get rid of FEFF marks in the generated files
+        result_file = regex.sub("\\.pdf$", ".html", new_file_path)
+        if(os.path.exists(result_file)):
+            return_code = subprocess.call([
+                'sed', '-i', 's/\\xff//;s/\\xfe//', result_file
+            ])
+
         os.remove(new_file_path)
         return return_code == 0
